move CartItem component into separate file & use mui icons, cleanup cartSlice imports
